kusto_scripts/data_prep/llm_as_judge/requirements.txt
# GPT-5.2 Soft Label Teacher Dependencies
# =========================================

# OpenAI SDK (includes async Azure client)
openai>=1.30.0

# Azure Authentication (for Key Vault access)
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0

# Azure Blob Storage (for reading input data)
azure-storage-blob>=12.19.0

# Tokenizer (for resolving token IDs for logit_bias)
tiktoken>=0.7.0

# Retry logic with exponential backoff
tenacity>=8.2.0

# Progress bars (async support via tqdm.asyncio)
tqdm>=4.66.0

# Optional: for calibration plots
# matplotlib>=3.7.0
# numpy>=1.24.0
